+++ lib/chat/db/copying.ex
@@ -6,31 +6,53 @@ defmodule Chat.Db.Copying do
 
   import Chat.Db.WriteQueue.ReadStream
 
-  alias Chat.Db.ChangeTracker
   alias Chat.Db.Common
+  alias Chat.Db.Copying.Progress
+  alias Chat.Db.Scope.Full, as: FullScope
   alias Chat.Db.WriteQueue
-  alias Chat.FileFs
 
-  def await_copied(from, to, keys \\ nil) do
+  def await_copied(from, to, keys_set \\ nil) do
     to_queue = Common.names(to, :queue)
-
-    stream = stream(from, to, nil, keys)
-
+    stream = stream(from, to, nil, keys_set)
     stream_keys = read_stream(stream, :keys)
+
     log_copying(from, to, stream_keys)
 
-    awaiter =
-      Task.async(fn ->
-        stream_keys
-        |> ChangeTracker.await_many(:timer.hours(1))
-      end)
+    stream |> WriteQueue.put_stream(to_queue)
 
-    stream
-    |> read_stream(awaiter: awaiter.pid)
-    |> WriteQueue.put_stream(to_queue)
+    Progress.new(stream_keys, to)
+    |> ensure_complete()
+    |> case do
+      :done -> :ok
+      {:stuck, progress} -> await_copied(from, to, Progress.get_unwritten_keys(progress))
+    end
+  end
 
-    Task.await(awaiter, :infinity)
-    to |> CubDB.file_sync()
+  defp ensure_complete(prev_progress, started? \\ false, stuck_for_ms \\ 0) do
+    prev_count = Progress.left_keys(prev_progress)
+    progress = Progress.eliminate_written(prev_progress)
+    # {time, progress} = :timer.tc(fn -> Progress.eliminate_written(prev_progress) end)
+    # time |> IO.inspect(label: "time")
+    count = Progress.left_keys(progress)
+
+    cond do
+      Progress.complete?(progress) ->
+        :done
+
+      stuck_for_ms > 100_000 ->
+        {:stuck, progress}
+
+      true ->
+        no_change = count == prev_count
+        delay = Progress.recheck_delay_in_ms(progress)
+        Process.sleep(delay)
+
+        ensure_complete(
+          progress,
+          started? or !no_change,
+          (started? && no_change && stuck_for_ms + delay) || 0
+        )
+    end
   end
 
   defp log_copying(from, to, keys) do
@@ -49,12 +71,12 @@ defmodule Chat.Db.Copying do
     |> Logger.info()
   end
 
-  defp stream(from, to, awaiter, keys)
+  defp stream(from, to, awaiter, keys_set)
 
   defp stream(from, to, awaiter, nil) do
     [from, to]
     |> Stream.map(fn db ->
-      Task.async(fn -> get_data_keys_set(db) end)
+      Task.async(fn -> FullScope.keys(db) end)
     end)
     |> Enum.to_list()
     |> Task.await_many(:timer.hours(1))
@@ -68,8 +90,8 @@ defmodule Chat.Db.Copying do
     end)
   end
 
-  defp stream(from, to, awaiter, src) do
-    dst = get_data_keys_set(to)
+  defp stream(from, to, awaiter, %MapSet{} = src) do
+    dst = FullScope.keys(to)
 
     keys =
       src
@@ -79,84 +101,7 @@ defmodule Chat.Db.Copying do
     read_stream(keys: keys, db: from, awaiter: awaiter)
   end
 
-  def get_data_keys_set(db) do
-    CubDB.with_snapshot(db, fn snap ->
-      {snap, MapSet.new()}
-      |> before_change_tracking()
-      |> after_change_tracking_till_chunk_keys()
-      |> chunk_keys()
-      |> after_chunk_keys_till_file_chunks()
-      |> after_file_chunks()
-      |> elem(1)
-    end)
-    |> join_fs_keys(db)
+  defp stream(from, to, awaiter, keys_list) when is_list(keys_list) do
+    stream(from, to, awaiter, MapSet.new(keys_list))
   end
-
-  defp chunk_keys({_snap, _set} = keys) do
-    keys |> join_keys_of(min_key: {:chunk_key, nil}, max_key: {:chunk_key, ""})
-  end
-
-  defp before_change_tracking(keys) do
-    keys |> join_keys_of(max_key: {:change_tracking_marker, 0}, max_key_inclusive: false)
-  end
-
-  defp after_change_tracking_till_chunk_keys(keys) do
-    keys
-    |> join_keys_of(
-      min_key: {:"change_tracking_marker\0", 0},
-      max_key: {:chunk_key, 0},
-      max_key_inclusive: false
-    )
-  end
-
-  defp after_chunk_keys_till_file_chunks(keys) do
-    keys
-    |> join_keys_of(
-      min_key: {:chunk_key, ""},
-      max_key: {:file_chunk, nil, nil, nil},
-      max_key_inclusive: false
-    )
-  end
-
-  defp after_file_chunks(keys) do
-    keys
-    |> join_keys_of(min_key: {:"file_chunk\0", 0, 0, 0})
-  end
-
-  defp join_keys_of({snap, set}, select_opts) do
-    CubDB.Snapshot.select(snap, select_opts)
-    |> Stream.map(fn {k, _v} -> k end)
-    |> MapSet.new()
-    |> MapSet.union(set)
-    |> then(&{snap, &1})
-  end
-
-  defp join_fs_keys(keys, db) do
-    db
-    |> CubDB.data_dir()
-    |> then(&"#{&1}_files")
-    |> FileFs.relative_filenames()
-    |> Enum.map(&filename_to_chunk_key/1)
-    |> Enum.reject(&is_nil/1)
-    |> MapSet.new()
-    |> MapSet.union(keys)
-  end
-
-  defp filename_to_chunk_key(<<
-         _::binary-size(3),
-         hash::binary-size(64),
-         ?/,
-         start::binary-size(20),
-         ?/,
-         finish::binary-size(20)
-       >>) do
-    {
-      :file_chunk,
-      hash |> Base.decode16!(case: :lower),
-      start |> String.to_integer(),
-      finish |> String.to_integer()
-    }
-  end
-
-  defp filename_to_chunk_key(_), do: nil
 end