+++ CHANGELOG.md
@@ -12,12 +12,15 @@ straightforward as possible.
 
 ### Added
 - extend cargo settings (camera sensors && weight sensor)
+- concurrent copying tracking
 
 ### Changed
 - more copying and changetracker logging
 
 ### Fixed
 - uploader background breaking the page when leaving room 
+- copying stuck
+- backup stuck on ejecting drive while copying
 
 
 ## 2023-05-22_3fefe32___2023-05-24_98fba24